Test(tutorial6) : Fixed the error with types difference

+++ tutorial6/src/main.rs
@@ -1,4 +1,4 @@
 fn main() {
-    let cond = 2 < 3.4;
+    let cond = (2 as f32) < 3.4;
     println!("{}", cond)
 }
